+++ AsBuiltReport.Veeam.VBR/Src/Private/Diagram/Get-AbrVbrHAClusterInfo.ps1
@@ -6,7 +6,7 @@ function Get-AbrHAClusterInfo {
     .DESCRIPTION
         Build a diagram of the configuration of Veeam VBR in PDF/PNG/SVG formats using Psgraph.
     .NOTES
-        Version:        1.0.0
+        Version:        1.0.1
         Author:         Jonathan Colon
         Twitter:        @jcolonfzenpr
         Github:         rebelinux
@@ -21,7 +21,7 @@ function Get-AbrHAClusterInfo {
         try {
             Write-PScriboMessage "Collecting HA Cluster Node information from $($VBRServer)."
 
-            $HACluster = Get-VBRHighAvailabilityCluster
+            $HACluster = Get-VBRHighAvailabilityCluster -WarningAction SilentlyContinue
 
             if ($HACluster) {
                 $HAClusterNodeInfo = @()
@@ -41,8 +41,8 @@ function Get-AbrHAClusterInfo {
 
                     $HAClusterNodeInfo += [PSCustomObject]@{
                         Name = $Node.Hostname
-                        Label = Add-NodeIcon -Name "$($Node.Hostname)" -IconType 'VBR_Server' -Align 'Center' -RowsOrdered $Rows -ImagesObj $Images -IconDebug $IconDebug -FontSize 18 -FontBold -TableBackgroundColor $BackupServerBGColor -CellBackgroundColor $BackupServerBGColor -FontColor $Fontcolor
-                        Spacer = Add-NodeIcon -Name ' ' -IconType 'VBR_Bid_Arrow' -Align 'Center' -ImagesObj $Images -IconDebug $IconDebug -TableBackgroundColor $BackupServerBGColor -CellBackgroundColor $BackupServerBGColor -FontColor $Fontcolor
+                        Label = Add-NodeIcon -Name "$($Node.Hostname)" -IconType 'VBR_Server' -Align 'Center' -RowsOrdered $Rows -ImagesObj $Images -IconDebug $IconDebug -FontSize 18 -FontBold -TableBackgroundColor $MainGraphBGColor -FontColor $Fontcolor
+                        Spacer = Add-NodeImage -Name 'Database Snapshots' -ImagesObj $Images -IconType 'VBR_Bid_Arrow' -IconDebug $IconDebug -TableBackgroundColor $MainGraphBGColor -IconPath $IconPath
                         Role = $Node.Role
                     }
                 }
@@ -52,9 +52,9 @@ function Get-AbrHAClusterInfo {
                     'Cluster IP' = $HACluster.ClusterEndpoint
                 }
 
-                $DNSNode = Add-NodeIcon -Name 'DNS Server' -IconType 'VBR_Tape_Drive' -Align 'Center' -ImagesObj $Images -IconDebug $IconDebug -FontSize 18 -FontBold -TableBackgroundColor $MainGraphBGColor -CellBackgroundColor $BackupServerBGColor -FontColor $Fontcolor -TableLayout Vertical
+                $DNSNode = Add-NodeIcon -Name 'DNS Server' -IconType 'VBR_Tape_Drive' -Align 'Center' -ImagesObj $Images -IconDebug $IconDebug -FontSize 18 -FontBold -TableBackgroundColor $MainGraphBGColor -FontColor $Fontcolor -TableLayout Vertical
 
-                $EndpointNode = Add-NodeIcon -Name $HACluster.ClusterDnsName.split('.')[0] -IconType 'VBR_GrayArrow' -Align 'Left' -ImagesObj $Images -IconDebug $IconDebug -FontSize 18 -FontBold -TableBackgroundColor $MainGraphBGColor -CellBackgroundColor $BackupServerBGColor -FontColor $Fontcolor -TableLayout Horizontal -AditionalInfo $EndPointTable -IconPath $IconPath
+                $EndpointNode = Add-NodeIcon -Name $HACluster.ClusterDnsName.ToUpper().split('.')[0] -IconType 'VBR_AGENT_CSV_Logo' -Align 'Left' -ImagesObj $Images -IconDebug $IconDebug -FontSize 18 -FontBold -TableBackgroundColor $MainGraphBGColor -FontColor $Fontcolor -TableLayout Horizontal -AditionalInfo $EndPointTable -IconPath $IconPath
 
 
                 return [PSCustomObject]@{